Trading-system/requirements.txt
ccxt==4.2.85
pandas==2.1.4
numpy==1.26.3
python-telegram-bot==20.7
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
sqlite3